still some refactor to make development later easier
